requirements.txt
# Data Engineering
pyspark==3.5.3

# Web Framework
fastapi==0.115.6
uvicorn==0.34.0

# Database
pymongo==4.10.1

# Validation
pydantic==2.10.6
email-validator==2.2.0

# Auth
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Config
python-dotenv==1.0.1

# Testing
pytest==8.3.4
pytest-mock==3.14.0
httpx==0.28.1
mongomock==4.2.0.post1

# Semantic Search (M3)
openai>=1.60.0
openai-agents>=0.1.0
qdrant-client>=1.12.0

# Future: Azure Blob Storage (uncomment for M2+)
# azure-storage-blob==12.24.0
